Move columniterator to tables
